requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fcm-django==2.0.1
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
install==1.3.5
msgpack==1.0.8
promise==2.3
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
rebase==1.3.2
requests==2.31.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
